+++ pkg/streamingpromql/functions_test.go
@@ -98,8 +98,10 @@ func TestFunctionDeduplicateAndMerge(t *testing.T) {
 		"sgn":                          `sgn({__name__=~"float.*"})`,
 		"sin":                          `sin({__name__=~"float.*"})`,
 		"sinh":                         `sinh({__name__=~"float.*"})`,
-		"sort":                         `<skip>`, // sort() and sort_desc() don't drop the metric name, so this test doesn't apply.
-		"sort_desc":                    `<skip>`, // sort() and sort_desc() don't drop the metric name, so this test doesn't apply.
+		"sort":                         `<skip>`, // sort*() functions don't drop the metric name, so this test doesn't apply.
+		"sort_by_label":                `<skip>`, // sort*() functions don't drop the metric name, so this test doesn't apply.
+		"sort_by_label_desc":           `<skip>`, // sort*() functions don't drop the metric name, so this test doesn't apply.
+		"sort_desc":                    `<skip>`, // sort*() functions don't drop the metric name, so this test doesn't apply.
 		"sqrt":                         `sqrt({__name__=~"float.*"})`,
 		"stddev_over_time":             `stddev_over_time({__name__=~"float.*"}[1m])`,
 		"stdvar_over_time":             `stdvar_over_time({__name__=~"float.*"}[1m])`,